Move tests to separate files
also test that async tests are successfully Send (currently they are not)
