+++ crates/swss-common/Cargo.toml
@@ -22,5 +22,6 @@ tokio = { version = "1", optional = true, features = ["net"] }
 bindgen = "0.70.1"
 
 [dev-dependencies]
+paste = "1.0.15"
 rand = "0.8.5"
-tokio = { version = "1", features = ["rt", "macros", "time"] }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "time"] }
